teste do crud feno e racao

+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/dtos/EntradaFenoRacaoResponseDTO.java
@@ -48,7 +48,7 @@ public record EntradaFenoRacaoResponseDTO(
 
         SituacaoLoteFenoRacao situacao,
 
-        Integer usuarioCadastroId,
+        Long usuarioCadastroId,
 
         LocalDateTime dataCadastro
