+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/dtos/SolicitacaoTransferenciaFenoRacaoResponseDTO.java
@@ -41,13 +41,13 @@ public record SolicitacaoTransferenciaFenoRacaoResponseDTO(
 
         StatusSolicitacaoTransferenciaFenoRacao status,
 
-        Integer solicitadoPorId,
+        Long solicitadoPorId,
 
         String solicitadoPorNome,
 
         LocalDateTime dataSolicitacao,
 
-        Integer respondidoPorId,
+        Long respondidoPorId,
 
         String respondidoPorNome,
