+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/dtos/TransferenciaFenoRacaoResponseDTO.java
@@ -43,7 +43,7 @@ public record TransferenciaFenoRacaoResponseDTO(
 
         Integer saldoPosteriorOrigem,
 
-        Integer aprovadoPorId,
+        Long aprovadoPorId,
 
         String aprovadoPorNome,
