+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/model/TransferenciaFenoRacao.java
@@ -117,7 +117,7 @@ public class TransferenciaFenoRacao {
             name = "aprovadoPorID",
             nullable = false
     )
-    private Integer aprovadoPorId;
+    private Long aprovadoPorId;
 
     @Column(
             name = "dataTransferencia",
