+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/repository/MovimentacaoFenoRacaoRepository.java
@@ -41,7 +41,7 @@ public interface MovimentacaoFenoRacaoRepository
 
     List<MovimentacaoFenoRacao>
     findByUsuarioIdOrderByDataOperacaoDescDataCadastroDesc(
-            Integer usuarioId
+            Long usuarioId
     );
 
     List<MovimentacaoFenoRacao>
