+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/repository/SolicitacaoTransferenciaFenoRacaoRepository.java
@@ -35,7 +35,7 @@ public interface SolicitacaoTransferenciaFenoRacaoRepository
 
     List<SolicitacaoTransferenciaFenoRacao>
     findBySolicitadoPorIdOrderByDataSolicitacaoDesc(
-            Integer solicitadoPorId
+            Long solicitadoPorId
     );
 
     Optional<SolicitacaoTransferenciaFenoRacao>
